backend/api/gemini_service.py
# Google Gemini Integration for RAG Chatbot
# Handles Google's Gemini API calls for generating responses based on retrieved content

import google.generativeai as genai
import os
import asyncio
from typing import List, Dict, Any, Optional
from pydantic import BaseModel
from dotenv import load_dotenv

load_dotenv()  # Load environment variables from .env file

class GeminiConfig:
    """Configuration for Gemini integration"""
    def __init__(self):
        self.api_key = os.getenv("GEMINI_API_KEY")
        self.model = os.getenv("GEMINI_MODEL", "gemini-pro")
        self.temperature = float(os.getenv("GEMINI_TEMPERATURE", "0.7"))
        self.max_tokens = int(os.getenv("GEMINI_MAX_TOKENS", "1000"))

        # Set the API key
        if self.api_key:
            genai.configure(api_key=self.api_key)
        else:
            raise ValueError("GEMINI_API_KEY environment variable not set")

class Message(BaseModel):
    role: str  # "user" or "model"
    content: str

class GeminiService:
    """Service class for interacting with Google's Gemini API"""

    def __init__(self):
        self.config = GeminiConfig()
        self.model = genai.GenerativeModel(self.config.model)

    def generate_embedding(self, text: str) -> List[float]:
        """Generate embedding for a text using Google's embedding API"""
        try:
            # Using Vertex AI or Google's embedding API
            # For now, using a placeholder - the actual implementation would require:
            # from vertexai.language_models import TextEmbeddingModel
            # embeddings = TextEmbeddingModel.from_pretrained("textembedding-gecko@001")
            # embedding = embeddings.get_embeddings([text])
            # return embedding[0].values
            print("Embedding generation requires Google Cloud Vertex AI or separate embedding service")
            # Placeholder implementation - in reality, you would use Google's embedding service
            # For now, return a dummy embedding of size 768 (common size for Google embeddings)
            return [0.0] * 768
        except Exception as e:
            print(f"Error generating embedding: {str(e)}")
            return []

    async def generate_embedding_async(self, text: str) -> List[float]:
        """Async version of generate_embedding"""
        try:
            # Placeholder for async implementation
            print("Async embedding generation requires Google Cloud Vertex AI or separate embedding service")
            return [0.0] * 768  # Placeholder
        except Exception as e:
            print(f"Error generating embedding: {str(e)}")
            return []

    def generate_response(self,
                         prompt: str,
                         context: Optional[str] = None,
                         temperature: Optional[float] = None) -> str:
        """Generate a response based on the provided prompt and optional context"""
        try:
            # Construct the full prompt with context if provided
            full_prompt = prompt
            if context:
                full_prompt = f"Context: {context}\n\nQuestion: {prompt}"

            # Generate content using the Gemini model
            response = self.model.generate_content(
                full_prompt,
                generation_config=genai.types.GenerationConfig(
                    temperature=temperature or self.config.temperature,
                    max_output_tokens=self.config.max_tokens
                )
            )

            return response.text
        except Exception as e:
            print(f"Error generating response: {str(e)}")
            return "Sorry, I encountered an error while processing your request."

    async def generate_response_async(self,
                                    prompt: str,
                                    context: Optional[str] = None,
                                    temperature: Optional[float] = None) -> str:
        """Async version of generate_response"""
        try:
            # For now, using sync method in async wrapper as the Gemini API
            # doesn't have true async support in the basic SDK
            return self.generate_response(prompt, context, temperature)
        except Exception as e:
            print(f"Error generating response: {str(e)}")
            return "Sorry, I encountered an error while processing your request."

    def validate_response_accuracy(self,
                                  response: str,
                                  source_chunks: List[Dict[str, Any]]) -> Dict[str, Any]:
        """Validate that the response is consistent with the source material"""
        # This is a simplified validation - in practice, this would be more sophisticated
        validation_result = {
            "is_consistent": True,
            "confidence_score": 0.85,
            "factual_errors": [],
            "hallucination_detected": False,
            "sources_used": [chunk['id'] for chunk in source_chunks]
        }

        # In a real implementation, we would:
        # 1. Check if the response contains information not present in the sources
        # 2. Compare key facts in the response to the source material
        # 3. Use additional LLM calls to verify accuracy

        return validation_result

# Global instance of Gemini service
gemini_service = GeminiService()