backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-frontmatter==1.0.0
qdrant-client==1.11.0
google-generativeai==0.4.1
pydantic-settings==2.1.0
numpy==2.1.0